Перенес из main.js вспомогательные функции в util.js, а код необходимый для формирования данных фотоописаный в data.js, настроен экспорт и импорт
